python/requirements.txt
# Python 3.8
# pip 21.1.1

Flask==1.1.2
requests==2.25.1
Pillow==8.2.0
numpy==1.20.3
pandas==1.2.4
scikit-learn==0.24.2
torch==1.8.1
torchvision==0.9.1
